// src/types/gallery.type.ts
// Generated by https://quicktype.io

export interface GetGalleryResponse {
  data: Datum[];
  linked: any[];
  fields: Field[];
  status: Status;
}

export interface Datum {
  id: number;
  mgalId: number;
  mgalName: string;
  mgalType: string;
  mgalUrl: string;
  mgalStatus: number;
  links: Links;
}

export interface Links {}

export interface Field {
  name: string;
  label: string;
  size: number;
  type: Type;
}

export enum Type {
  Datetime = 'datetime',
  Integer = 'integer',
  String = 'string',
}

export interface Status {
  page: number;
  totalRecords: number;
}